requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==1.0.0
    # via
    #   tensorboard
    #   tensorflow
astunparse==1.6.3
    # via tensorflow
attrs==21.4.0
    # via
    #   flake8-bugbear
    #   flake8-eradicate
    #   pytest
cachetools==5.0.0
    # via google-auth
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.2
    # via pip-tools
cvl-hdf5 @ git+https://gitlab.com/CVisionLab/cvl-python-hdf5.git@v1.0.0
    # via -r requirements.in
cycler==0.11.0
    # via matplotlib
eradicate==2.1.0
    # via flake8-eradicate
flake8==4.0.1
    # via
    #   -r requirements.in
    #   flake8-broken-line
    #   flake8-bugbear
    #   flake8-comprehensions
    #   flake8-docstrings
    #   flake8-eradicate
    #   flake8-gl-codeclimate
    #   flake8-polyfill
    #   flake8-print
    #   flake8-quotes
    #   flake8-tuple
    #   pep8-naming
flake8-broken-line==0.4.0
    # via -r requirements.in
flake8-bugbear==22.3.23
    # via -r requirements.in
flake8-comprehensions==3.8.0
    # via -r requirements.in
flake8-docstrings==1.6.0
    # via -r requirements.in
flake8-eradicate==1.2.0
    # via -r requirements.in
flake8-gl-codeclimate==0.1.5
    # via -r requirements.in
flake8-import-order==0.18.1
    # via -r requirements.in
flake8-polyfill==1.0.2
    # via pep8-naming
flake8-print==4.0.0
    # via -r requirements.in
flake8-quotes==3.3.1
    # via -r requirements.in
flake8-tuple==0.4.1
    # via -r requirements.in
flatbuffers==2.0
    # via tensorflow
fonttools==4.31.2
    # via matplotlib
gast==0.4.0
    # via tensorflow
gdcm==1.1
    # via -r requirements.in
google-auth==2.6.2
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.44.0
    # via
    #   tensorboard
    #   tensorflow
h5py==3.2.0
    # via
    #   -r requirements.in
    #   cvl-hdf5
    #   tensorflow
idna==3.3
    # via requests
imageio==2.16.1
    # via
    #   imgaug
    #   scikit-image
imgaug==0.4.0
    # via -r requirements.in
importlib-metadata==4.11.3
    # via markdown
iniconfig==1.1.1
    # via pytest
keras==2.7.0
    # via tensorflow
keras-preprocessing==1.1.2
    # via tensorflow
kiwisolver==1.4.2
    # via matplotlib
libclang==13.0.0
    # via tensorflow
loguru==0.6.0
    # via -r requirements.in
markdown==3.3.6
    # via tensorboard
matplotlib==3.5.1
    # via
    #   -r requirements.in
    #   imgaug
    #   seaborn
mccabe==0.6.1
    # via flake8
mypy==0.942
    # via -r requirements.in
mypy-extensions==0.4.3
    # via mypy
networkx==2.7.1
    # via scikit-image
nibabel==3.2.2
    # via -r requirements.in
numpy==1.22.3
    # via
    #   -r requirements.in
    #   cvl-hdf5
    #   h5py
    #   imageio
    #   imgaug
    #   keras-preprocessing
    #   matplotlib
    #   nibabel
    #   opencv-python
    #   opt-einsum
    #   pandas
    #   pylibjpeg
    #   pylibjpeg-libjpeg
    #   pylibjpeg-openjpeg
    #   pylibjpeg-rle
    #   pywavelets
    #   scikit-image
    #   scipy
    #   seaborn
    #   tensorboard
    #   tensorflow
    #   tifffile
oauthlib==3.2.0
    # via requests-oauthlib
opencv-python==4.5.5.64
    # via imgaug
opt-einsum==3.3.0
    # via tensorflow
packaging==21.3
    # via
    #   matplotlib
    #   nibabel
    #   pytest
    #   scikit-image
pandas==1.4.1
    # via seaborn
pep517==0.12.0
    # via pip-tools
pep8-naming==0.12.1
    # via -r requirements.in
pillow==9.0.1
    # via
    #   imageio
    #   imgaug
    #   matplotlib
    #   scikit-image
pip-tools==6.6.0
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
protobuf==3.19.4
    # via
    #   tensorboard
    #   tensorflow
py==1.11.0
    # via pytest
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycodestyle==2.8.0
    # via
    #   flake8
    #   flake8-import-order
    #   flake8-print
pydicom==2.3.0
    # via -r requirements.in
pydocstyle==6.1.1
    # via flake8-docstrings
pyflakes==2.4.0
    # via flake8
pylibjpeg==1.3.0
    # via -r requirements.in
pylibjpeg-libjpeg==1.1.0
    # via -r requirements.in
pylibjpeg-openjpeg==1.1.1
    # via
    #   -r requirements.in
    #   pylibjpeg
pylibjpeg-rle==1.3.0
    # via pylibjpeg
pyparsing==3.0.7
    # via
    #   matplotlib
    #   packaging
pytest==7.1.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-magic==0.4.25
    # via -r requirements.in
pytz==2022.1
    # via pandas
pywavelets==1.3.0
    # via scikit-image
requests==2.27.1
    # via
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.8
    # via google-auth
scikit-image==0.19.2
    # via
    #   -r requirements.in
    #   imgaug
scipy==1.8.0
    # via
    #   imgaug
    #   scikit-image
    #   seaborn
seaborn==0.11.2
    # via -r requirements.in
shapely==1.8.1.post1
    # via imgaug
six==1.16.0
    # via
    #   absl-py
    #   astunparse
    #   flake8-print
    #   flake8-tuple
    #   google-auth
    #   google-pasta
    #   grpcio
    #   imgaug
    #   keras-preprocessing
    #   python-dateutil
    #   tensorflow
snowballstemmer==2.2.0
    # via pydocstyle
tensorboard==2.8.0
    # via tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow==2.7.0
    # via -r requirements.in
tensorflow-addons==0.16.1
    # via -r requirements.in
tensorflow-estimator==2.7.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.24.0
    # via tensorflow
termcolor==1.1.0
    # via tensorflow
tifffile==2022.3.25
    # via scikit-image
tomli==2.0.1
    # via
    #   mypy
    #   pep517
    #   pytest
tqdm==4.63.1
    # via -r requirements.in
typeguard==2.13.3
    # via tensorflow-addons
typing-extensions==4.1.1
    # via
    #   mypy
    #   tensorflow
urllib3==1.26.9
    # via requests
werkzeug==2.1.0
    # via tensorboard
wheel==0.37.1
    # via
    #   astunparse
    #   pip-tools
    #   tensorboard
    #   tensorflow
wrapt==1.14.0
    # via tensorflow
zipp==3.7.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
